requirements.txt
langchain-openai
langchain-together
pydantic
openai
together
python-dotenv